Fix up database creation and password-reset logic
Previously, the case where a database exists but the user doesn't was
handled by the "reset password" job.  Since switching to SRCFLib, this
should be done with the "create user and database" job (which will skip
either of the two steps as needed).

We also tend to link users to the database password reset page directly,
which doesn't work if they don't have an account yet, so this change
introduces redirects to the opposite of create/reset when appropriate.